// src/types/options.ts
export interface OptionsFile {
  endianness: "little" | "big";
  musicVolume: number;
  soundVolume: number;
  gamma: number;
  gameSensitivity: number;
  interfaceSensitivity: number;
  interfaceOpacity: number;
  difficulty: number;
  viewBobbing: boolean;
  invertLook: boolean;
  southpaw: boolean;
  verticalSplitscreen: boolean;
  inGameGamertags: boolean;
  autosaveTimer: number;
  splitscreenGamertags: boolean;
  hints: boolean;
  inGameTooltips: boolean;
  renderClouds: boolean;
  displayHud: boolean;
  displayHand: boolean;
  customSkinAnimation: boolean;
  deathMessages: boolean;
  hudSize: number;
  hudSizeSplitscreen: number;
  animatedCharacter: boolean;
  classicCrafting: boolean;
  caveSounds: boolean;
  gameChat: boolean;
  minecartSounds: boolean;
  displayGameMessages: boolean;
  displaySaveIcon: boolean;
  flyingViewRolling: boolean;
  showGlideGhost: boolean;
  showGlideGhostPath: boolean;
  autoJump: boolean;
  chosenSkin: number;
  playerCape: number;
  favoriteSkins: number[];
  actions: {
    jump: number;
    use: number;
    action: number;
    cycleHeldItemLeft: number;
    cycleHeldItemRight: number;
    inventory: number;
    drop: number;
    sneakDismount: number;
    crafting: number;
    changeCameraMode: number;
    flyLeft: number;
    flyRight: number;
    flyUp: number;
    flyDown: number;
    sprint: number;
    pickBlock: number;
    previousPlayer: number;
    nextPlayer: number;
    spectateNoise: number;
    cancelSpectating: number;
    confirmReady: number;
    vote: number;
    restartSection: number;
    restartRace: number;
    lookBehind: number;
  };

  rawData: Uint8Array;
}